pyproject.toml
[project]
name = "nonebot-plugin-amrita"
version = "0.1.0"
description = "AmritaCore support for NoneBot2"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "amrita-core>=0.7.1",
    "nonebot-plugin-orm>=0.8.3",
    "nonebot2>=2.4.4",
]

[project.urls]
"Homepage" = "https://github.com/LiteSuggarDEV/nonebot_plugin_amrita"
"Source" = "https://github.com/LiteSuggarDEV/nonebot_plugin_amrita"
"Issue Tracker" = "https://github.com/LiteSuggarDEV/nonebot_plugin_amrita/issues"


[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "F",     # Pyflakes
    "W",     # pycodestyle warnings
    "E",     # pycodestyle errors
    "UP",    # pyupgrade
    "ASYNC", # flake8-async
    "C4",    # flake8-comprehensions
    "T10",   # flake8-debugger
    "PYI",   # flake8-pyi
    "PT",    # flake8-pytest-style
    "Q",     # flake8-quotes
    "RUF",   # Ruff-specific rules
    "I",     # isort
    "PERF",  # pylint-performance
]
ignore = [
    "E402",   # module-import-not-at-top-of-file
    "E501",   # line-too-long
    "UP037",  # quoted-annotation
    "RUF001", # ambiguous-unicode-character-string
    "RUF002", # ambiguous-unicode-character-docstring
    "RUF003", # ambiguous-unicode-character-comment
]

[tool.pyright]
typeCheckingMode = "standard"

[tool.nonebot]
plugins=[
    "nonebot_plugin_amrita",
    "nonebot_plugin_localstore",
    "nonebot_plugin_orm",
]

[dependency-groups]
dev = [
    "nonebot2[fastapi]>=2.4.4",
    "nonebot-plugin-orm[default]>=0.8.3",
]
